requirements.txt
python-json-config
azure-storage-blob